constructora/requirements.txt
# requirements.txt - Dependencias del proyecto Flask Python
# Sistema de Gestión de Constructora - Solo Python + HTML + CSS

# Framework web Flask
Flask==3.0.0
Werkzeug==3.0.1

# Conexión a base de datos PostgreSQL
psycopg2-binary==2.9.9

# Variables de entorno
python-dotenv==1.0.0

# Seguridad - Hash de contraseñas
bcrypt==4.1.2

# Utilidades adicionales (opcionales pero recomendadas)
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0

# Para desarrollo (opcional)
# Estas dependencias son solo para desarrollo y testing
# flask-debug-toolbar==0.13.1
pytest==7.4.3
pytest-flask==1.3.0